cloudguard/pyproject.toml
[project]
name = "cloudguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-environ>=0.13.0",
    "django-ninja>=1.5.3",
    "gunicorn>=25.1.0",
    "kubernetes>=35.0.0",
    "postgres>=4.0",
    "psycopg>=3.3.3",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--nomigrations" # Para testes unitários serem instantâneos
